+++ container/container_test.go
@@ -1143,7 +1143,10 @@ func TestKeyboard(t *testing.T) {
 					testcanvas.MustNew(image.Rect(20, 10, 40, 20)),
 					&widgetapi.Meta{Focused: true},
 					widgetapi.Options{WantKeyboard: widgetapi.KeyScopeFocused},
-					&terminalapi.Keyboard{Key: keyboard.KeyEnter},
+					&fakewidget.Event{
+						Ev:   &terminalapi.Keyboard{Key: keyboard.KeyEnter},
+						Meta: &widgetapi.EventMeta{Focused: true},
+					},
 				)
 				return ft
 			},
@@ -1188,7 +1191,10 @@ func TestKeyboard(t *testing.T) {
 					testcanvas.MustNew(image.Rect(0, 0, 20, 20)),
 					&widgetapi.Meta{},
 					widgetapi.Options{WantKeyboard: widgetapi.KeyScopeGlobal},
-					&terminalapi.Keyboard{Key: keyboard.KeyEnter},
+					&fakewidget.Event{
+						Ev:   &terminalapi.Keyboard{Key: keyboard.KeyEnter},
+						Meta: &widgetapi.EventMeta{},
+					},
 				)
 
 				// Widget that isn't focused and only wants focused events.
@@ -1205,7 +1211,10 @@ func TestKeyboard(t *testing.T) {
 					testcanvas.MustNew(image.Rect(20, 10, 40, 20)),
 					&widgetapi.Meta{Focused: true},
 					widgetapi.Options{WantKeyboard: widgetapi.KeyScopeFocused},
-					&terminalapi.Keyboard{Key: keyboard.KeyEnter},
+					&fakewidget.Event{
+						Ev:   &terminalapi.Keyboard{Key: keyboard.KeyEnter},
+						Meta: &widgetapi.EventMeta{Focused: true},
+					},
 				)
 				return ft
 			},
@@ -1412,7 +1421,6 @@ func TestMouse(t *testing.T) {
 					testcanvas.MustNew(image.Rect(25, 10, 50, 20)),
 					&widgetapi.Meta{},
 					widgetapi.Options{WantMouse: widgetapi.MouseScopeWidget},
-					&terminalapi.Keyboard{},
 				)
 
 				// The target widget receives the mouse event.
@@ -1421,8 +1429,14 @@ func TestMouse(t *testing.T) {
 					testcanvas.MustNew(image.Rect(25, 0, 50, 10)),
 					&widgetapi.Meta{Focused: true},
 					widgetapi.Options{WantMouse: widgetapi.MouseScopeWidget},
-					&terminalapi.Mouse{Position: image.Point{24, 9}, Button: mouse.ButtonLeft},
-					&terminalapi.Mouse{Position: image.Point{24, 9}, Button: mouse.ButtonRelease},
+					&fakewidget.Event{
+						Ev:   &terminalapi.Mouse{Position: image.Point{24, 9}, Button: mouse.ButtonLeft},
+						Meta: &widgetapi.EventMeta{},
+					},
+					&fakewidget.Event{
+						Ev:   &terminalapi.Mouse{Position: image.Point{24, 9}, Button: mouse.ButtonRelease},
+						Meta: &widgetapi.EventMeta{Focused: true},
+					},
 				)
 				return ft
 			},
@@ -1495,7 +1509,6 @@ func TestMouse(t *testing.T) {
 					testcanvas.MustNew(image.Rect(25, 10, 50, 20)),
 					&widgetapi.Meta{},
 					widgetapi.Options{WantMouse: widgetapi.MouseScopeWidget},
-					&terminalapi.Keyboard{},
 				)
 
 				// The target widget receives the mouse event.
@@ -1504,14 +1517,20 @@ func TestMouse(t *testing.T) {
 					testcanvas.MustNew(image.Rect(26, 1, 49, 9)),
 					&widgetapi.Meta{Focused: true},
 					widgetapi.Options{WantMouse: widgetapi.MouseScopeWidget},
-					&terminalapi.Mouse{Position: image.Point{22, 7}, Button: mouse.ButtonLeft},
-					&terminalapi.Mouse{Position: image.Point{22, 7}, Button: mouse.ButtonRelease},
+					&fakewidget.Event{
+						Ev:   &terminalapi.Mouse{Position: image.Point{22, 7}, Button: mouse.ButtonLeft},
+						Meta: &widgetapi.EventMeta{},
+					},
+					&fakewidget.Event{
+						Ev:   &terminalapi.Mouse{Position: image.Point{22, 7}, Button: mouse.ButtonRelease},
+						Meta: &widgetapi.EventMeta{Focused: true},
+					},
 				)
 				return ft
 			},
 		},
 		{
-			desc:     "key event focuses the next container, widget with KeyScopeFocused does not get the key as it was not focused yet",
+			desc:     "key event focuses the next container, widget with KeyScopeFocused gets the key as it is now focused",
 			termSize: image.Point{50, 20},
 			container: func(ft *faketerm.Terminal) (*Container, error) {
 				c, err := New(
@@ -1542,6 +1561,10 @@ func TestMouse(t *testing.T) {
 					testcanvas.MustNew(image.Rect(0, 0, 25, 20)),
 					&widgetapi.Meta{Focused: true},
 					widgetapi.Options{WantKeyboard: widgetapi.KeyScopeFocused},
+					&fakewidget.Event{
+						Ev:   &terminalapi.Keyboard{Key: keyboard.KeyTab},
+						Meta: &widgetapi.EventMeta{Focused: true},
+					},
 				)
 				fakewidget.MustDraw(
 					ft,
@@ -1591,6 +1614,10 @@ func TestMouse(t *testing.T) {
 					testcanvas.MustNew(image.Rect(25, 0, 50, 20)),
 					&widgetapi.Meta{Focused: true},
 					widgetapi.Options{WantKeyboard: widgetapi.KeyScopeFocused},
+					&fakewidget.Event{
+						Ev:   &terminalapi.Keyboard{Key: keyboard.KeyTab},
+						Meta: &widgetapi.EventMeta{Focused: true},
+					},
 				)
 				return ft
 			},
@@ -1627,19 +1654,26 @@ func TestMouse(t *testing.T) {
 					testcanvas.MustNew(image.Rect(0, 0, 25, 20)),
 					&widgetapi.Meta{Focused: true},
 					widgetapi.Options{WantKeyboard: widgetapi.KeyScopeFocused},
+					&fakewidget.Event{
+						Ev:   &terminalapi.Keyboard{Key: keyboard.KeyTab},
+						Meta: &widgetapi.EventMeta{Focused: true},
+					},
 				)
 				fakewidget.MustDraw(
 					ft,
 					testcanvas.MustNew(image.Rect(25, 0, 50, 20)),
 					&widgetapi.Meta{},
 					widgetapi.Options{WantKeyboard: widgetapi.KeyScopeGlobal},
-					&terminalapi.Keyboard{Key: keyboard.KeyTab},
+					&fakewidget.Event{
+						Ev:   &terminalapi.Keyboard{Key: keyboard.KeyTab},
+						Meta: &widgetapi.EventMeta{},
+					},
 				)
 				return ft
 			},
 		},
 		{
-			desc:     "key event moves focus from a widget with KeyScopeFocused, the originally focused widget gets the key",
+			desc:     "key event moves focus from a widget with KeyScopeFocused, the newly focused widget gets the key",
 			termSize: image.Point{50, 20},
 			container: func(ft *faketerm.Terminal) (*Container, error) {
 				c, err := New(
@@ -1673,13 +1707,22 @@ func TestMouse(t *testing.T) {
 					testcanvas.MustNew(image.Rect(0, 0, 25, 20)),
 					&widgetapi.Meta{},
 					widgetapi.Options{WantKeyboard: widgetapi.KeyScopeFocused},
-					&terminalapi.Keyboard{Key: keyboard.KeyTab},
+					// Also gets the key, since we are sending two events above.
+					&fakewidget.Event{
+						Ev: &terminalapi.Keyboard{Key: keyboard.KeyTab},
+						// Also is focused at the time of the first event.
+						Meta: &widgetapi.EventMeta{Focused: true},
+					},
 				)
 				fakewidget.MustDraw(
 					ft,
 					testcanvas.MustNew(image.Rect(25, 0, 50, 20)),
 					&widgetapi.Meta{Focused: true},
 					widgetapi.Options{WantKeyboard: widgetapi.KeyScopeFocused},
+					&fakewidget.Event{
+						Ev:   &terminalapi.Keyboard{Key: keyboard.KeyTab},
+						Meta: &widgetapi.EventMeta{Focused: true},
+					},
 				)
 				return ft
 			},
@@ -1745,7 +1788,7 @@ func TestMouse(t *testing.T) {
 		},
 		{
 			desc:     "MouseScopeContainer, event forwarded if it falls on the container's border",
-			termSize: image.Point{21, 20},
+			termSize: image.Point{23, 20},
 			container: func(ft *faketerm.Terminal) (*Container, error) {
 				return New(
 					ft,
@@ -1771,17 +1814,20 @@ func TestMouse(t *testing.T) {
 
 				fakewidget.MustDraw(
 					ft,
-					testcanvas.MustNew(image.Rect(1, 1, 20, 19)),
+					testcanvas.MustNew(image.Rect(1, 1, 22, 19)),
 					&widgetapi.Meta{Focused: true},
 					widgetapi.Options{WantMouse: widgetapi.MouseScopeWidget},
-					&terminalapi.Mouse{Position: image.Point{-1, -1}, Button: mouse.ButtonLeft},
+					&fakewidget.Event{
+						Ev:   &terminalapi.Mouse{Position: image.Point{-1, -1}, Button: mouse.ButtonLeft},
+						Meta: &widgetapi.EventMeta{Focused: true},
+					},
 				)
 				return ft
 			},
 		},
 		{
 			desc:     "MouseScopeGlobal, event forwarded if it falls on the container's border",
-			termSize: image.Point{21, 20},
+			termSize: image.Point{23, 20},
 			container: func(ft *faketerm.Terminal) (*Container, error) {
 				return New(
 					ft,
@@ -1807,10 +1853,13 @@ func TestMouse(t *testing.T) {
 
 				fakewidget.MustDraw(
 					ft,
-					testcanvas.MustNew(image.Rect(1, 1, 20, 19)),
+					testcanvas.MustNew(image.Rect(1, 1, 22, 19)),
 					&widgetapi.Meta{Focused: true},
 					widgetapi.Options{WantMouse: widgetapi.MouseScopeWidget},
-					&terminalapi.Mouse{Position: image.Point{-1, -1}, Button: mouse.ButtonLeft},
+					&fakewidget.Event{
+						Ev:   &terminalapi.Mouse{Position: image.Point{-1, -1}, Button: mouse.ButtonLeft},
+						Meta: &widgetapi.EventMeta{Focused: true},
+					},
 				)
 				return ft
 			},
@@ -1848,7 +1897,7 @@ func TestMouse(t *testing.T) {
 		},
 		{
 			desc:     "MouseScopeContainer event forwarded if it falls outside of widget's canvas",
-			termSize: image.Point{20, 20},
+			termSize: image.Point{22, 20},
 			container: func(ft *faketerm.Terminal) (*Container, error) {
 				return New(
 					ft,
@@ -1870,17 +1919,20 @@ func TestMouse(t *testing.T) {
 
 				fakewidget.MustDraw(
 					ft,
-					testcanvas.MustNew(image.Rect(0, 5, 20, 15)),
+					testcanvas.MustNew(image.Rect(0, 4, 22, 15)),
 					&widgetapi.Meta{Focused: true},
 					widgetapi.Options{WantMouse: widgetapi.MouseScopeWidget},
-					&terminalapi.Mouse{Position: image.Point{-1, -1}, Button: mouse.ButtonLeft},
+					&fakewidget.Event{
+						Ev:   &terminalapi.Mouse{Position: image.Point{-1, -1}, Button: mouse.ButtonLeft},
+						Meta: &widgetapi.EventMeta{Focused: true},
+					},
 				)
 				return ft
 			},
 		},
 		{
 			desc:     "MouseScopeGlobal event forwarded if it falls outside of widget's canvas",
-			termSize: image.Point{20, 20},
+			termSize: image.Point{22, 20},
 			container: func(ft *faketerm.Terminal) (*Container, error) {
 				return New(
 					ft,
@@ -1902,10 +1954,13 @@ func TestMouse(t *testing.T) {
 
 				fakewidget.MustDraw(
 					ft,
-					testcanvas.MustNew(image.Rect(0, 5, 20, 15)),
+					testcanvas.MustNew(image.Rect(0, 4, 22, 15)),
 					&widgetapi.Meta{Focused: true},
 					widgetapi.Options{WantMouse: widgetapi.MouseScopeWidget},
-					&terminalapi.Mouse{Position: image.Point{-1, -1}, Button: mouse.ButtonLeft},
+					&fakewidget.Event{
+						Ev:   &terminalapi.Mouse{Position: image.Point{-1, -1}, Button: mouse.ButtonLeft},
+						Meta: &widgetapi.EventMeta{Focused: true},
+					},
 				)
 				return ft
 			},
@@ -2014,7 +2069,10 @@ func TestMouse(t *testing.T) {
 					testcanvas.MustNew(image.Rect(0, 10, 20, 20)),
 					&widgetapi.Meta{},
 					widgetapi.Options{WantMouse: widgetapi.MouseScopeWidget},
-					&terminalapi.Mouse{Position: image.Point{-1, -1}, Button: mouse.ButtonLeft},
+					&fakewidget.Event{
+						Ev:   &terminalapi.Mouse{Position: image.Point{-1, -1}, Button: mouse.ButtonLeft},
+						Meta: &widgetapi.EventMeta{},
+					},
 				)
 				return ft
 			},
@@ -2046,14 +2104,17 @@ func TestMouse(t *testing.T) {
 					testcanvas.MustNew(image.Rect(0, 5, 20, 15)),
 					&widgetapi.Meta{Focused: true},
 					widgetapi.Options{WantMouse: widgetapi.MouseScopeWidget},
-					&terminalapi.Mouse{Position: image.Point{0, 0}, Button: mouse.ButtonLeft},
+					&fakewidget.Event{
+						Ev:   &terminalapi.Mouse{Position: image.Point{0, 0}, Button: mouse.ButtonLeft},
+						Meta: &widgetapi.EventMeta{Focused: true},
+					},
 				)
 				return ft
 			},
 		},
 		{
 			desc:     "mouse position adjusted relative to widget's canvas, horizontal offset",
-			termSize: image.Point{30, 20},
+			termSize: image.Point{40, 30},
 			container: func(ft *faketerm.Terminal) (*Container, error) {
 				return New(
 					ft,
@@ -2075,10 +2136,13 @@ func TestMouse(t *testing.T) {
 
 				fakewidget.MustDraw(
 					ft,
-					testcanvas.MustNew(image.Rect(6, 0, 24, 20)),
+					testcanvas.MustNew(image.Rect(6, 0, 33, 30)),
 					&widgetapi.Meta{Focused: true},
 					widgetapi.Options{WantMouse: widgetapi.MouseScopeWidget},
-					&terminalapi.Mouse{Position: image.Point{0, 0}, Button: mouse.ButtonLeft},
+					&fakewidget.Event{
+						Ev:   &terminalapi.Mouse{Position: image.Point{0, 0}, Button: mouse.ButtonLeft},
+						Meta: &widgetapi.EventMeta{Focused: true},
+					},
 				)
 				return ft
 			},
@@ -2481,7 +2545,7 @@ func TestUpdate(t *testing.T) {
 		},
 		{
 			desc:     "newly placed widget gets keyboard events",
-			termSize: image.Point{10, 10},
+			termSize: image.Point{12, 10},
 			container: func(ft *faketerm.Terminal) (*Container, error) {
 				return New(
 					ft,
@@ -2510,7 +2574,10 @@ func TestUpdate(t *testing.T) {
 					cvs,
 					&widgetapi.Meta{Focused: true},
 					widgetapi.Options{WantKeyboard: widgetapi.KeyScopeFocused},
-					&terminalapi.Keyboard{Key: keyboard.KeyEnter},
+					&fakewidget.Event{
+						Ev:   &terminalapi.Keyboard{Key: keyboard.KeyEnter},
+						Meta: &widgetapi.EventMeta{Focused: true},
+					},
 				)
 				testcanvas.MustApply(cvs, ft)
 				return ft
@@ -2518,7 +2585,7 @@ func TestUpdate(t *testing.T) {
 		},
 		{
 			desc:     "newly placed widget gets mouse events",
-			termSize: image.Point{20, 10},
+			termSize: image.Point{22, 10},
 			container: func(ft *faketerm.Terminal) (*Container, error) {
 				return New(
 					ft,
@@ -2542,7 +2609,10 @@ func TestUpdate(t *testing.T) {
 					cvs,
 					&widgetapi.Meta{Focused: true},
 					widgetapi.Options{WantMouse: widgetapi.MouseScopeWidget},
-					&terminalapi.Mouse{Position: image.Point{0, 0}, Button: mouse.ButtonRelease},
+					&fakewidget.Event{
+						Ev:   &terminalapi.Mouse{Position: image.Point{0, 0}, Button: mouse.ButtonRelease},
+						Meta: &widgetapi.EventMeta{Focused: true},
+					},
 				)
 				testcanvas.MustApply(cvs, ft)
 				return ft