+++ widgets/textinput/textinput.go
@@ -267,7 +267,7 @@ func (ti *TextInput) keyboard(k *terminalapi.Keyboard) (bool, string) {
 
 // Keyboard processes keyboard events.
 // Implements widgetapi.Widget.Keyboard.
-func (ti *TextInput) Keyboard(k *terminalapi.Keyboard) error {
+func (ti *TextInput) Keyboard(k *terminalapi.Keyboard, meta *widgetapi.EventMeta) error {
 	if submitted, text := ti.keyboard(k); submitted {
 		// Mutex must be released when calling the callback.
 		// Users might call container methods from the callback like the
@@ -279,7 +279,7 @@ func (ti *TextInput) Keyboard(k *terminalapi.Keyboard) error {
 
 // Mouse processes mouse events.
 // Implements widgetapi.Widget.Mouse.
-func (ti *TextInput) Mouse(m *terminalapi.Mouse) error {
+func (ti *TextInput) Mouse(m *terminalapi.Mouse, meta *widgetapi.EventMeta) error {
 	ti.mu.Lock()
 	defer ti.mu.Unlock()